[stable @ 1267655304e736d6bc78d4ddac7ec0261a017f55]
Implemented a usage limit over the amount of memory projected 
for use by the coLinux VM instances. The current limit is 
'host_mem - 64MB' for machines with host_mem above and equal 
to 256MB, and host_mem*3/4 for machines with less memory.

The usage limit is set below the total amount of memory 
in order to prevent memory depletion which could have led 
to breakage in the host OS.